Security, file manager and static
Added a skeleton for a file manager
Added a security component which will be responsible for user
authentication, XSS securing strings etc.
Introduced the concept of static methods (using the Users.cfc) to my
framework.

Since the Recipe.cfc, Comment.cfc and User.cfc model actual objects and
thus heavily resemble each other I have decided to focus on the User.cfc
for now. At this stage it is ready for unit testing and once that is
done and I am satisfied with how it works I will replicate the changes
across to the other cfc's