Login and authentication
Logging in and authentication of users is now in place.
This is also where the first controller is added to the system.
Much of scripts and markup in Login.cfm is temporary in-line for
development and testing reasons. It will be moved to the proper files
once the functionality is solidified.