Login and authentication changes again
Once again the login routine has been changed.
Had a hell of a time getting the sessions to work in the right
application scope. In order to get OnRequestStart to not fire on the
login logic I had to separate the login page into its own application.
Then I couldn't get cflogin to log in to the correct application despite
the use of applicationToken targeting the main app. The solution I came
up with was to give both apps the same name which ended up having the
double benefit of me being able to set session variables from within the
login application that can be accessed by the main app.