Final login/authentication touch
Although I am not convinced (after doing some research) that my login
code is following best practices I am satisfied with its robustness and
security for now. Otherwise I can keep on tweaking it forever and never
work on the rest.

My biggest bug bear is still that I current have both Application.cfc's
sharing the same name as I can't figure out how to keep the login app
separate, have it log my user in to the recipedb app while also making
sure that the authentication check on each request is working correctly.
I will have a look at that another day.

In addition to this there are minor fixes and tweaks here and there to
the rest of the code. Mostly still work-in-progress on using a
controller to generate a view (UserSettings.cfm) and all that jazz.