Add recipe and duplicate check
The major addition is the Add recipe-page which also has a built-in
duplicates check.
Alongside that are many smaller tweaks, optimizations and fixes.
The Recipe-view has also been prepped for use but the mechanism for
relocating the view to it after adding a new recipe has not yet to be
implemented.
Another change of note is splitting out methods related to instance
modelling into Model.cfc which the other relevant CFCs inherit.