Comment model preparation
Mostly changes to Comment.cfc to bring it up to standard and ready for
integration.
Some minor fixes for the Recipe.cfc
Changed how the duplicate recipe check works, basically passing the data
to the view in a different format so that the view and the model are
fully decoupled.