Salted passwords, model fixes
Main change has been the introduction of salted passwords plus the hash
strength has been upped to SHA-512.
The duplicates recipe view had a bug with an unclosed quote causing
issues and the controller had a few fixes as well to accomodate the
changes I made last time.
User.cfc and Recipe.cfc both had changes to streamline how create()
works and also just to bring them on the same page.
The Model.cfc had a comment added so it's easier to remember what needs
to be changed when a new db column is added.
Main.js had a bug fix that prevented you from being able to log in and
CreateDB.sql was amended to include the new password salt column and
some changes to other db column lengths .