Toolbox, db check-up script
Added the beginnings of a dev/admin toolkit. The idea came after I
experimented with cfdbinfo to create a scrip that could do a "health"
inspection of the database. That script is in place now and is very
handy during the development work.

The rest of the toolbox will probably consist of scripts for making new
users, altering the database and also for initiating unit tests, which I
will start writing shortly