Recipe list styling and tweaks, plus prep work
Most of this commit is about the recipe list. Styling has been
implemented, a few tweaks to the method for getting the data and I also
rearranged the columns and the data.

A lot of the other stuff is architecture that I want to implement some
day:
An ajax proxy to make the ajax calls to the CFCs secure
System.cfc got changed although I am still unsure what to use it for and
how
There's been a few things commented out to I want to redesign
Comments have been put into places where I still have to make something
or decide how to do it