More preparations and minor fixes
Mostly more preparations. Some static objects are being put in the
(semi)permanent scopes. Which reminds me I'd love to get rid of the
double Application.cfc's but that's going to require some rejigging that
I'd rather spend the energy on later.

Guess the most noteworthy changes are:
Some prep work for editing the recipe itself has been done, which will
be my main focus going forward.
Changed the models to return a copy of themselves when using init,
rather than this, as that fits the static strategy better.