Massive update
There's a ton of changes in this update:

- I've done a bit of cleanup for the Application.cfc's, putting some
static components in the application-scope and trying to clean up the
logout/not logged in-logic.

- Added a bunch of tools to the toolbox. One set is for manipulating
users and the other is for a new set of features, mainly handling
patchnotes and the roadmap of upcoming features.

- The recipes are now editable. It's not final, but it works. Decided to
use an AJAX submission for ALL fields once editing has been enabled.

- There are numerous smaller fixes and changes to style and JS. I still
want to change the overall structure of the app to use page-relocations
rather than the SPA-style of injecting views into the main page, but
that comes after the main features have been implemented.

- The three communication-views (stats, roadmap and patch notes) are
making use of the AjaxProxy. The rest of the app will be updated to use
it later on.