Authentication final fix
Finally managed to fix the double application.cfc mess that was caused
by me not being able to find a decent, non-derpy way of dealing with
login and authentication. Rather than having the authentication
process(es) in their own separate cfc I now have it all inside of
onRequestStart() instead. Still not 100% sure of the way I am doing it
now is "the best" but it works and it feels cleaner, and a lot more in
line with what I have seen on the net. I know you can not always trust
the net and must be happy with your own code, but it is tough when you
have no experience with authentication in CF.