Various changes, moving closer to release
User settings can now be saved, yay!
The dumps in Login.cfm and Main.cfm have been commented out.
Commented out Favorites and Statistics from the menu as they'll be
implemented later.
Removed the caching of the query for the recipe list.
Added a Toolbar-header to the toolbar when viewing a Recipe.
Various fixes and smaller changes.