Various fixes
Fixed the listFind() oversight in Application.cfc for the toolbox
access.
Removed word-break CSS rule from the recipe sections.
Removed the ingredients-column from the recipe list.
Added the missing logic for actually saving the displayname when adding
a new user.
Added some plugins to TinyMCE init and also enabled forced paste as
text.
Added a bunch of useful cleaning methods to the SecurityManager for
future use.