Change password, style changes and JS overhaul
A tonne of changes:
- The biggest functional change is the addition of a change password
function
- Lots of smaller style changes, additions of new styles but mostly
adjustments to make bring some things inline with the general style idea
and to make it scale better on tablets
- JS has had large changes, mostly focused on overhauling existing
stuff. In this case the notification system has been been the target,
trying to generalize it by adding functions to show the different kinds
of notifications, so that I could pull all of that piecemeal code out of
all the page specific JS.
- Made a few fixes, such as ensuring you'd get a proper error message if
you tried to access a non-existing vacancy by fiddling with the url
variable. Also made it so that the index page doesn't show you the
initial 1666 date when you log in as a new user, but instead welcomes
you as a new user.
- Put in some initial groundwork for working with modals, with a few JS
methods and some CSS rules.