Security and control/error-handling
A bunch of bugs have been fixed but this round of changes has really
focused on improving the control and error flow.

- Models methods that don't return data are now all void. Error handling
in the models in basically to throw exceptions whenever things don't
work the way they should.

- Controller methods ALL return a structure with statuscode and data.
Still have to refine the details.

- Had a look at all the notifications and changed their positions
(fixed, top or bottom) so that it makes more sense.

- Added "variables" in front of all calls to self inside the models.

- Added some extra security to the AjaxProxy in the form of a list of
CFCs that are allowed to be called.

- Added "Published" to the recipe list as a column. Also added a JS
render function to ensure it is shown as "yes/no" but still filters
correctly on 1 and 0.

- Changed the way published-status is updated in the DOM via JS. It now
gets the new status as an argument from the AJAX callback instead of
reacting to a data-attribute or class-attribute.

- Added a method for showing a LOADING-notification for when you
navigate pages. Currently disabled as I am not sure if I want to use it.

- Brushed up the onJavascriptError() JS method a bit. Now shows raw
trace when debug mode is on.