File uploads!
Pictures can now be uploaded for use with the recipes. That's pretty much all there is to this change, and it touches a lot of files. It also marks the first use of a dialog/pop-up and I may change how that works later on to streamline it a bit better.

There are also a few other changes bundled with this:

- Biggest one is that the Notes folder has been removed. The patch notes, roadmap - together with the location of the pictures - has now been changed to be dependent on a config file in xml form that lives inside the Assets-folder.

- A bit of house cleaning has been done. Test-files have been removed, unused pictures have been removed and I made some minor fixes to the toolbox.