Streamlining dialog procedures
Made some changes to our first and so far only dialog. Dialog JS is now put inside RecipeDB.dialog and when closed (by invoking main.removeDialog) that object is cleared so that no dialog specific code lingers in memory. This also makes it ready for the next dialog, in case we have multiple on the page.