Tools update and some fixes
Apparently a user agent string can be way longer than 200 chars so increased it to 300.
Wrapped the login page in try/catch so we can display neater error messages to users if something goes wrong.
Updated the patchnotes and roadmap tools so they use the new pathing from the previous update.