AJAX fileupload rework plus more
Biggest change is that fileuploads via AJAX has been reworked and now works in a manner I like better. Basically it's now possible to upload files via AJAX going via the AjaxProxy and calling another CFC. Before it seemed AJAX fileuploads were only possible when calling directly to CFM-controllers.

The HTMLHead has been moved to Modules and all references updated (hopefully).

All actions (delete, update and change picture) for recipes have ownership checks added to the controllers so that you can't fiddle with the JS and manipulate other recipes.

AjaxProxy statuscodes have been changed to 10x to distinguish them from controller statuscodes. This is mostly relevant for when debugging.

The FileManager has had delete-methods added to them.  When you delete a recipe the picture is also deleted and if uploading a picture fails, the temp picture is deleted.

The delete-method for Models now clears itself so that the instance is completely unusable. Only there for sanity but nice none the less.